rename and move shared test files to utils folder
