--- cypress/support/_utils/index.js
@@ -1,5 +1,0 @@
-import { testBreadcrumbs } from "./testBreadcrumbs";
-import { testMenus } from "./testMenus";
-import { testPagination } from "./testPagination";
-
-export { testBreadcrumbs, testMenus, testPagination };
