+++ cypress/support/utils/index.js
@@ -1,0 +1,13 @@
+import { testBreadcrumbs } from "./testBreadcrumbs";
+import { testMenus } from "./testMenus";
+import { testPagination } from "./testPagination";
+import { search } from "./testSearch";
+import { fillRequestVisitForm } from "./testRequestVisitForm";
+
+export {
+  testBreadcrumbs,
+  testMenus,
+  testPagination,
+  search,
+  fillRequestVisitForm,
+};
